DTO-地址簿的增删改重筑，id都改为了long,可能引发问题

+++ src/main/java/com/sftc/web/dao/mybatis/TokenMapper.java
@@ -12,7 +12,7 @@ public interface TokenMapper {
 
     Token getToken(int order_id);
 
-    Token getTokenById(int user_id);
+    Token getTokenById(long user_id);
 
     Token getTokenByMobile(String mobile);
